Split out Resource & ResourceQuerySet
